+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2018-07-31
+
 - New query parser
 - Added support for preparsed queries
 - Enum type fixes
@@ -34,7 +36,8 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/ajk/specialist-server/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/ajk/specialist-server/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/ajk/specialist-server/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/ajk/specialist-server/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/ajk/specialist-server/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/ajk/specialist-server/compare/0.1.0...0.2.0